requirements.txt
Django==4.1.3
python-dotenv==0.21.0
psycopg2-binary==2.9.5
pytest-django==4.5.2